+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         tools:ignore="UnusedAttribute, GoogleAppIndexingWarning">
 
         <activity
-            android:name="net.chineseguide.jukuu.ui.MainActivity"
+            android:name="net.chineseguide.jukuu.ui.main.MainActivity"
             android:label="@string/app_name"
             android:windowSoftInputMode="adjustPan|stateHidden">
             <intent-filter>
