+++ app/src/main/java/net/chineseguide/jukuu/domain/entity/AppLocale.kt
@@ -1,6 +1,7 @@
 package net.chineseguide.jukuu.domain.entity
 
 enum class AppLocale(val id: Int, val languageCode: String, val countryCode: String) {
+
     ENGLISH(0,"en", ""),
     RUSSIAN(1,"ru", ""),
     CHINESE(2, "zh", ""),
