+++ app/src/main/java/net/chineseguide/jukuu/domain/repository/SentenceRepository.kt
@@ -1,0 +1,10 @@
+package net.chineseguide.jukuu.domain.repository
+
+import net.chineseguide.jukuu.domain.entity.Sentence
+
+interface SentenceRepository {
+
+    fun get(query: String): List<Sentence>
+
+    fun getNext(query: String, sentenceIndex: Int): List<Sentence>
+}
