+++ app/src/main/java/net/chineseguide/jukuu/domain/usecase/GetNextSentencesUseCase.kt
@@ -1,7 +1,7 @@
 package net.chineseguide.jukuu.domain.usecase
 
-import net.chineseguide.jukuu.data.repository.SentenceRepository
 import net.chineseguide.jukuu.domain.entity.Sentence
+import net.chineseguide.jukuu.domain.repository.SentenceRepository
 import javax.inject.Inject
 
 class GetNextSentencesUseCase @Inject constructor(
