+++ app/src/main/java/net/chineseguide/jukuu/ui/App.kt
@@ -1,3 +1,5 @@
 package net.chineseguide.jukuu.ui
 
+import net.chineseguide.jukuu.ui.base.BaseApp
+
 class App : BaseApp()
