+++ app/src/main/java/net/chineseguide/jukuu/ui/main/MainActivity.kt
@@ -1,5 +1,6 @@
-package net.chineseguide.jukuu.ui
+package net.chineseguide.jukuu.ui.main
 
 import net.chineseguide.jukuu.R
+import net.chineseguide.jukuu.ui.base.BaseActivity
 
 class MainActivity : BaseActivity(R.layout.activity_main)
