+++ app/src/main/java/net/chineseguide/jukuu/ui/util/extensions/SearchViewExtensions.kt
@@ -1,4 +1,4 @@
-package net.chineseguide.jukuu.ui.util
+package net.chineseguide.jukuu.ui.util.extensions
 
 import androidx.appcompat.widget.SearchView
