+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/mainFragment"
-    android:name="net.chineseguide.jukuu.ui.MainFragment"
+    android:name="net.chineseguide.jukuu.ui.main.MainFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_constraintBottom_toBottomOf="parent"
